OneDrive/Рабочий стол/TESTiKi 2/app/deps.py
from fastapi import Depends, HTTPException, status
from jose import JWTError, jwt
from app.database import get_db
from app.models import User
from sqlalchemy.ext.asyncio import AsyncSession
from app.schemas import TokenData
from typing import Optional
import os

SECRET_KEY = os.getenv("SECRET_KEY", "supersecretkey")
ALGORITHM = "HS256"

async def get_current_user(token: str = Depends(lambda: None), db: AsyncSession = Depends(get_db)) -> User:
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    if token is None:
        raise credentials_exception
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id: Optional[int] = payload.get("sub")
        if user_id is None:
            raise credentials_exception
        token_data = TokenData(user_id=int(user_id))
    except JWTError:
        raise credentials_exception
    user = await db.get(User, token_data.user_id)
    if user is None:
        raise credentials_exception
    return user 