OneDrive/Рабочий стол/TESTiKi 2/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy[asyncio]
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
PyJWT
pytest
httpx
email-validator
python-multipart 